Make association type a long instead of an int. Switch order of association type and recipient for associations. This is done to make it consistent with the statement tx.

+++ lang/shared/src/main/scala/com/ltonetwork/lang/v1/traits/Domain.scala
@@ -34,8 +34,8 @@ object Tx {
   case class MassTransfer(p: Proven, transferCount: Long, totalAmount: Long, transfers: IndexedSeq[TransferItem], attachment: ByteVector) extends Tx
   case class Anchor(p: Proven, anchorCount: Long, anchors: IndexedSeq[ByteVector])                                                        extends Tx
   case class Data(p: Proven, data: IndexedSeq[DataItem[_]])                                                                               extends Tx
-  case class IssueAssociation(p: Proven, assocType: Int, recipient: LangAddress)                                                          extends Tx
-  case class RevokeAssociation(p: Proven, assocType: Int, recipient: LangAddress)                                                         extends Tx
+  case class IssueAssociation(p: Proven, assocType: Long, recipient: LangAddress)                                                         extends Tx
+  case class RevokeAssociation(p: Proven, assocType: Long, recipient: LangAddress)                                                        extends Tx
   case class Sponsorship(p: Proven, recipient: LangAddress)                                                                               extends Tx
   case class CancelSponsorship(p: Proven, recipient: LangAddress)                                                                         extends Tx
   case class Register(p: Proven, accounts: IndexedSeq[PublicKey])                                                                         extends Tx
